feat: id - Long 타입으로 변경

See merge request s08-webmobile1-sub2/S08P12B206!47

+++ backend/src/main/java/coogether/backend/domain/UserJoinList.java
@@ -35,7 +35,7 @@ public class UserJoinList {
     @Id
     @GeneratedValue
     @Column(name = "user_join_list_id", nullable = false)
-    private int userJoinListId;
+    private Long userJoinListId;
 
     @CreatedDate // 참여자 입장 시간
     @Column(name = "user_join_reg_time", updatable = false, nullable = false)
