+++ backend/src/main/java/coogether/backend/dto/IngredientListDto.java
@@ -14,7 +14,7 @@ import javax.persistence.ManyToOne;
 
 @Data
 public class IngredientListDto {
-    private int ingredientListId;
+    private Long ingredientListId;
     private String ingredientAmount;
 
     /////////////////////
