+++ backend/src/main/java/coogether/backend/repository/ingredientlist/IngredientListRepository.java
@@ -8,8 +8,8 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface IngredientListRepository extends JpaRepository<IngredientList, Integer> {
+public interface IngredientListRepository extends JpaRepository<IngredientList, Long> {
 
     @Query("select il from IngredientList il where il.recipe.recipeId = :recipeId ")
-    List<IngredientList> findByRecipeName(@Param("recipeId") int recipeId);
+    List<IngredientList> findByRecipeName(@Param("recipeId") Long recipeId);
 }
